Code Refactor: 1. MC moves refactor 2. MC_MOVE class: Every move is divided into different stages: preparation, calculation, and acceptance.    * Adding modifications made easier 3. small optimization for CBMC. 5% speed up Examples: 1. Re-calculated examples: output are saved to output.txt in each example folder.    * As long as order of random numbers are preserved, simulation should give the same result everywhere, and correctness is gaurenteed 2. Allegro/LCLin ML potential example revisited, LCLin small bug fix

+++ src_clean/print_statistics.cuh
@@ -600,7 +600,7 @@ static inline void Gather_Averages_double(std::vector<double>& Array, double val
 ///////////////////////////////////////////////////////
 // Wrapper for the functions for printing statistics //
 ///////////////////////////////////////////////////////
-static inline void PrintAllStatistics(Components& SystemComponents, Simulations& Sims, size_t Cycles, int SimulationMode, size_t BlockAverageSize, Units& Constants)
+static inline void PrintAllStatistics(Components& SystemComponents, Simulations& Sims, size_t Cycles, int SimulationMode, int BlockAverageSize, Units& Constants)
 {
   for(size_t comp = 0; comp < SystemComponents.NComponents.x; comp++)
   {
